renamed to datastructures.ts and minor clean up
